Convert float32 to float values
This is probably a sign that my mocked return values are bad. Running
the application with real NetCDF diag files caused JSON serialization
errors because the mean and std deviation were all float32, not float,
so Python couldn't serialize them.

In any case I've tested the production image by hand to confirm that I
can get back the JSON from the NetCDF file. We should probably consider
setting up some integration tests with test data stored in S3.

+++ services/api/src/unified_graphics/diag.py
@@ -24,12 +24,12 @@ def get_diagnostics(loop: MinimLoop) -> Dict:
     obs_minus_fcast = ds["Obs_Minus_Forecast_adjusted"].values
 
     obs_count = len(obs_minus_fcast)
-    mean = np.mean(obs_minus_fcast)
-    std = np.std(obs_minus_fcast)
+    mean = float(np.mean(obs_minus_fcast))
+    std = float(np.std(obs_minus_fcast))
     counts, bins = np.histogram(obs_minus_fcast, bins="auto")
 
     distribution = [
-        {"lower": bins[idx], "upper": bins[idx + 1], "value": int(count)}
+        {"lower": float(bins[idx]), "upper": float(bins[idx + 1]), "value": int(count)}
         for idx, count in enumerate(counts)
     ]
